Encapsulated Theme, Onboarding as Rails::Engine and the admin is the core application
