+++ lib/ecrire/app/helpers/posts_helper.rb
@@ -4,4 +4,10 @@ module PostsHelper
 
     link_to t('posts.edit'), edit_admin_post_path(post.id), options
   end
+
+  def paginate(scope, options = {}, &block)
+    _with_routes Ecrire::Theme::Engine.routes do
+      super
+    end
+  end
 end
