+++ lib/ecrire/app/views/layouts/admin/application.html.erb
@@ -13,7 +13,7 @@
       <%= content_tag :nav, id: 'Menu', class: ['admin', controller_name] do %>
         <section>
           <div class='back'>
-            <%= link_to 'Home', :root %>
+            <%= link_to 'Home', '/' %>
           </div>
           <div class='posts'>
             <h4>
