--- lib/ecrire/tasks/routes.rake
@@ -1,7 +1,0 @@
-task routes: :environment do
-  Rails.application.paths.add 'config/routes.rb', with: 'config/routes.rb', eager_load: true
-  Rails.application.paths['config/routes.rb'].unshift *Ecrire::Railtie.paths['user:routes'].existent
-  Rails.application.paths['config/routes.rb'].existent.each do |path|
-    require path
-  end
-end
