+++ test/editor/controllers/sessions_controller_test.rb
@@ -2,7 +2,7 @@ class SessionsControllerTest < BaseControllerTest
   test 'should be able to log in' do
     post :create, session: {email: "pothibo@gmail.com", password: "123456789"}
     assert !proxy.user.nil?, "#{users(:pothibo).email} should be logged in"
-    assert_redirected_to root_path
+    assert_redirected_to '/'
   end
