+++ test/themes/template/routes.rb
@@ -1,4 +1,4 @@
-Ecrire::Application.routes.draw do
+Ecrire::Theme::Engine.routes.draw do
   get '/:year/:month/:id', controller: :posts, action: :show, constraints: { year: /\d+/, month: /\d+/ }, as: "post"
   get '/feed', controller: :posts, action: :index, defaults: { format: :rss }
